# engineering/contracts/http/openapi.json
{
  "components": {
    "schemas": {
...
      "CorrelationId": {
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
        "type": "string"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "$ref": "#/components/schemas/CorrelationId"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemDetailsError"
            },
            "type": "array"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "code",
          "correlationId",
          "status",
          "title",
          "type"
        ],
        "type": "object"
      },
      "ProblemDetailsError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "path"
        ],
        "type": "object"
      },
...
      "PropertyPhotoGalleryResponse": {
        "additionalProperties": false,
        "properties": {
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PropertyPhoto"
            },
            "type": "array"
          }
        },
        "required": [
          "photos"
        ],
        "type": "object"
      },
...
      "RegisterPropertyPhotoRequest": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "enum": [
              "BATHROOM_OR_SHOWER_ROOM",
              "BEDROOM_OR_SLEEPING_AREA",
              "BUILDING_EXTERIOR_OR_ENTRANCE",
              "KITCHEN_OR_KITCHENETTE",
              "LIVING_ROOM_OR_MAIN_ROOM",
              "MAIN_LIVING_SLEEPING_AREA",
              "OTHER"
            ],
            "type": "string"
          },
          "contentBase64": {
            "minLength": 1,
            "type": "string"
          },
          "contentType": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "type": "string"
          }
        },
        "required": [
          "category",
          "contentBase64",
          "contentType"
        ],
        "type": "object"
      },
...
      "UpdatePropertyCoreInformationRequest": {
        "additionalProperties": false,
        "properties": {
          "apartmentSubtype": {
            "enum": [
              "MULTI_ROOM",
              "STUDIO"
            ],
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "type": "string"
          },
          "location": {
            "additionalProperties": false,
            "properties": {
              "addressLine": {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              "city": {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              "country": {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              "district": {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              }
            },
            "required": [
              "addressLine",
              "city",
              "country",
              "district"
            ],
            "type": "object"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "location",
          "title"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned MonPiole HTTP transport contracts",
    "title": "MonPiole Public API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/properties/{propertyId}": {
      "get": {
        "operationId": "retrieveProperty",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Tenant-owned property",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid property identifier"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden property authority"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Property not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a tenant-owned property",
        "tags": [
          "Properties"
        ]
      },
      "put": {
        "operationId": "updatePropertyCoreInformation",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyCoreInformationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Property core information updated",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid property core information"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden property authority"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Property not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update property title, description and location",
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/v1/properties/{propertyId}/photos": {
      "get": {
        "operationId": "listPropertyPhotos",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPhotoGalleryResponse"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid identifier"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Property not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List available photos of a tenant-owned Property",
        "tags": [
          "Property photos"
        ]
      },
      "post": {
        "operationId": "registerPropertyPhoto",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPropertyPhotoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPhotoGalleryResponse"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid photo content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Property not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Persist photo content and register an available Property photo",
        "tags": [
          "Property photos"
        ]
      }
    },
    "/v1/properties/{propertyId}/photos/{photoId}": {
      "delete": {
        "operationId": "deletePropertyPhoto",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Non-primary photo deleted",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid identifier"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Property or photo not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Primary photo requires a replacement"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a non-primary Property photo",
        "tags": [
          "Property photos"
        ]
      }
    },
    "/v1/properties/{propertyId}/photos/{photoId}/content": {
      "get": {
        "operationId": "retrievePropertyPhotoContent",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Persisted image content",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Photo not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve persisted private Property photo content",
        "tags": [
          "Property photos"
        ]
      }
    },
    "/v1/properties/{propertyId}/photos/{photoId}/primary": {
      "put": {
        "operationId": "selectPropertyPrimaryPhoto",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "photoId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPhotoGalleryResponse"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid identifier"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Property or photo not found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Safe internal failure"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atomically select or replace the primary Property photo",
        "tags": [
          "Property photos"
        ]
      }
    },
...
  },
  "servers": [],
...
}